pyproject.toml
[tool.poetry]
name = "dataanalytics"
version = "0.1.0"
description = ""
authors = ["Alex <alexandr.basso@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
asgiref = "3.7.2"
build = "1.0.3"
cachecontrol = "0.13.1"
certifi = "2024.2.2"
charset-normalizer = "3.3.2"
cleo = "2.1.0"
colorama = "0.4.6"
contourpy = "1.2.0"
crashtest = "0.4.1"
cycler = "0.12.1"
defusedxml = "0.7.1"
diff-match-patch = "20230430"
distlib = "0.3.8"
django = "4.2"
django-import-export = "3.3.6"
dulwich = "0.21.7"
et-xmlfile = "1.1.0"
fastjsonschema = "2.19.1"
filelock = "3.13.1"
fonttools = "4.47.2"
idna = "3.6"
installer = "0.7.0"
jaraco-classes = "3.3.1"
keyring = "24.3.0"
kiwisolver = "1.4.5"
markuppy = "1.14"
matplotlib = "3.8.2"
more-itertools = "10.2.0"
msgpack = "1.0.7"
numpy = "1.26.3"
odfpy = "1.4.1"
openpyxl = "3.1.2"
packaging = "23.2"
pandas = "2.2.0"
pexpect = "4.9.0"
pillow = "10.2.0"
pkginfo = "1.9.6"
platformdirs = "3.11.0"
poetry = "1.7.1"
poetry-core = "1.8.1"
poetry-plugin-export = "1.6.0"
ptyprocess = "0.7.0"
pyparsing = "3.1.1"
pyproject-hooks = "1.0.0"
python-dateutil = "2.8.2"
python-dotenv = "1.0.1"
pytz = "2023.4"
pywin32-ctypes = "0.2.2"
pyyaml = "6.0.1"
rapidfuzz = "3.6.1"
requests = "2.31.0"
requests-toolbelt = "1.0.0"
setuptools = "68.2.2"
shellingham = "1.5.4"
six = "1.16.0"
sqlparse = "0.4.4"
tablib = "3.5.0"
tomlkit = "0.12.3"
trove-classifiers = "2024.1.31"
tzdata = "2023.4"
urllib3 = "2.2.0"
virtualenv = "20.25.0"
wheel = "0.41.2"
xlrd = "2.0.1"
xlwt = "1.3.0"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
